To help with final paper
"Analyzing The Effect of Noise" is a little discussion on what "main" results I have. 

"noise_analysis_workspace" is pure chaos. But it might have plots that are interesting. I tried labeling most of them as I went, but it was a workspace and I'm lazy, so I didn't. Let me know if there's something I can provide from there.